backend/requirements.txt
asgiref==3.11.1
cfgv==3.5.0
distlib==0.4.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
filelock==3.25.2
identify==2.6.18
nodeenv==1.10.0
platformdirs==4.9.4
pre_commit==4.5.1
PyJWT==2.12.1
python-discovery==1.2.1
python-dotenv==1.2.2
PyYAML==6.0.3
sqlparse==0.5.5
virtualenv==21.2.0
